pyproject.toml
[project]
name = "bot"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Nathyz", email = "swiftnathyz@gmail.com"},
]
dependencies = [
    "discord-py>=2.4.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.4.0",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "Apache-2.0"}


[tool.pdm]
